+++ include/monolish_generalized_eigen.hpp
@@ -5,12 +5,10 @@
 #include "./monolish_solver.hpp"
 #include "common/monolish_common.hpp"
 
-namespace monolish {
-
 /**
  * @brief handling eigenvalues and eigenvectors
  **/
-namespace generalized_eigen {
+namespace monolish::generalized_eigen {
 
 /**
  * @brief LOBPCG solver
@@ -97,5 +95,4 @@ public:
   std::string name() const { return "monolish::generalized_eigen::DC"; }
 };
 
-} // namespace generalized_eigen
-} // namespace monolish
+} // namespace monolish::generalized_eigen